+++ crates/wdl-lint/src/v1.rs
@@ -11,11 +11,13 @@ mod double_quotes;
 mod ending_newline;
 mod missing_runtime;
 mod no_curly_commands;
+mod snake_case;
 
 pub use double_quotes::*;
 pub use ending_newline::*;
 pub use missing_runtime::*;
 pub use no_curly_commands::*;
+pub use snake_case::*;
 
 /// A trait implemented by lint rules.
 pub trait Rule {
@@ -50,6 +52,7 @@ pub fn rules() -> Vec<Box<dyn Rule>> {
     let rules: Vec<Box<dyn Rule>> = vec![
         Box::new(DoubleQuotesRule),
         Box::new(NoCurlyCommandsRule),
+        Box::new(SnakeCaseRule),
         Box::new(MissingRuntimeRule),
         Box::new(EndingNewlineRule),
     ];